+++ framework/base/src/base/org/ofbiz/base/util/StringUtil.java
@@ -275,7 +275,7 @@ public class StringUtil {
     public static String cleanHexString(String str) {
         StringBuffer buf = new StringBuffer();
         for (int i = 0; i < str.length(); i++) {
-            if (str.charAt(i) != (int) 32 && str.charAt(i) != ':') {
+            if (str.charAt(i) != 32 && str.charAt(i) != ':') {
                 buf.append(str.charAt(i));
             }
         }
